GraphQL/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
httpx[http2]
python-dotenv
pydantic
aiocache
